Port dev-libs/openssl-1.0.1k to RISC-V
This port is actually super simple: we just tell OpenSSL that it's not
supposed to use any assembly code when building on RISC-V.  I'm not
entirely sure if this is because I'm just cross-compiling or if
OpenSSL defaults to AMD64 assembly, but this fixes the problem.

Note that while this nominally supports rv32, I didn't bother testing.