sys-libs/glibc: let default for src_prepare apply the patches
Remove the use of epatch to apply patches. In EAPI6, the default
behavior for src_prepare is to apply all the patches (defined in -p1
format) in the PATCHES bash array automatically.

Note that this changes the order in which patches are applied with
respect to user-supplied patches and actually corrects the
order. Previously, the epatch statement appeared after the call to
default, which applies the user-provided patches internally. This may
cause the later epatch statement to fail if it doesn't apply cleanly.

Package-Manager: Portage-2.3.24, Repoman-2.3.6
RepoMan-Options: --force